server: A new function "set_sd_defaults_from_token".
It sets the security descriptor along with a token that will be used to gather defaults, instead of always using the primary token.

Some objects take their defaults not from a primary token but from a different one (such as from the impersonation token or the process token).
This function can be used to create the various set_sd implementations for the objects that need it.
As a bonus, a NULL token will skip setting any defaults, this is useful for object implementations that would like to set their defaults _only_ upon creation.